Backshift Operator Apply the backshift operator or lag operator to a time series objective.

Backshift Operator

Apply the backshift operator or lag operator to a time series objective.

bshift(x, k = 1)

Arguments

x

univariate or multivariate time series.

k

number of lags.

Examples

x <- arima.sim(model = list(ar = 0.8, sd = 0.5), n = 120)
bshift(x, k = 12)
#>                [,1]
#>   [1,]           NA
#>   [2,]           NA
#>   [3,]           NA
#>   [4,]           NA
#>   [5,]           NA
#>   [6,]           NA
#>   [7,]           NA
#>   [8,]           NA
#>   [9,]           NA
#>  [10,]           NA
#>  [11,]           NA
#>  [12,]           NA
#>  [13,] -0.862519765
#>  [14,] -2.202415463
#>  [15,] -0.826569181
#>  [16,] -0.484766734
#>  [17,] -0.144127923
#>  [18,]  1.508246545
#>  [19,]  1.318635319
#>  [20,]  0.920911242
#>  [21,] -1.173358474
#>  [22,] -1.217924021
#>  [23,] -1.287785195
#>  [24,]  0.037079723
#>  [25,]  0.099698629
#>  [26,] -0.559364421
#>  [27,] -0.497456436
#>  [28,] -0.649448592
#>  [29,] -0.074761758
#>  [30,]  2.695608169
#>  [31,]  2.203017916
#>  [32,]  2.340123402
#>  [33,]  1.990293596
#>  [34,] -0.319485614
#>  [35,]  0.606497990
#>  [36,]  0.241961653
#>  [37,] -0.012517873
#>  [38,]  0.009163294
#>  [39,]  0.036891390
#>  [40,]  0.579340653
#>  [41,] -1.810642334
#>  [42,]  1.234043316
#>  [43,]  0.626013398
#>  [44,]  0.714166468
#>  [45,]  1.645679056
#>  [46,]  0.651454997
#>  [47,]  1.635116416
#>  [48,]  1.062196721
#>  [49,] -0.327805932
#>  [50,] -1.238095362
#>  [51,]  0.074581031
#>  [52,]  0.191335459
#>  [53,]  0.641697177
#>  [54,] -1.186092827
#>  [55,] -2.419610568
#>  [56,] -1.651538110
#>  [57,]  0.016089925
#>  [58,]  0.249568223
#>  [59,]  1.517947963
#>  [60,]  1.738268158
#>  [61,]  1.997362573
#>  [62,]  1.487954386
#>  [63,]  1.362545224
#>  [64,]  0.999708892
#>  [65,]  2.724110455
#>  [66,]  3.477681123
#>  [67,]  3.530936166
#>  [68,]  3.380973262
#>  [69,]  2.156521346
#>  [70,]  2.835751969
#>  [71,] -0.343732758
#>  [72,] -0.430679982
#>  [73,]  0.089345804
#>  [74,] -0.310474469
#>  [75,]  0.175808000
#>  [76,]  1.203748396
#>  [77,]  2.011711336
#>  [78,]  1.571266174
#>  [79,]  1.743161859
#>  [80,]  3.067412098
#>  [81,]  2.099568515
#>  [82,]  2.626002698
#>  [83,]  3.417628515
#>  [84,]  2.437462787
#>  [85,]  1.562756654
#>  [86,]  0.464772668
#>  [87,] -0.684918733
#>  [88,] -1.343476416
#>  [89,] -2.831056561
#>  [90,] -2.955383145
#>  [91,] -2.922848511
#>  [92,] -2.874942135
#>  [93,] -2.072826575
#>  [94,] -0.679806340
#>  [95,] -0.752727723
#>  [96,] -2.001592639
#>  [97,] -1.342736823
#>  [98,] -1.515988911
#>  [99,] -0.644191268
#> [100,]  1.611497445
#> [101,]  1.714056397
#> [102,] -0.313036415
#> [103,] -0.001027348
#> [104,]  1.072016374
#> [105,]  2.896982361
#> [106,]  2.767039667
#> [107,]  3.605445779
#> [108,]  3.310923170
#> [109,]  2.756322529
#> [110,]  2.227352756
#> [111,]  2.385493215
#> [112,]  1.645743999
#> [113,]  0.788331117
#> [114,]  0.822814316
#> [115,] -0.487948216
#> [116,]  0.455826092
#> [117,]  0.446380503
#> [118,] -0.948012608
#> [119,] -1.703322146
#> [120,] -0.908316123